+++ src/GRA.Controllers/ViewModel/Challenges/ChallengesListViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.ComponentModel;
 using GRA.Controllers.ViewModel.Shared;
 using Microsoft.AspNetCore.Mvc.Rendering;
 
@@ -10,11 +11,13 @@ namespace GRA.Controllers.ViewModel.Challenges
         public GRA.Domain.Model.ChallengeGroup ChallengeGroup { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
         public string Search { get; set; }
+        public int? Program { get; set; }
         public string Categories { get; set; }
         public bool? Favorites { get; set; }
         public bool IsActive { get; set; }
         public bool IsLoggedIn { get; set; }
 
+        public SelectList ProgramList { get; set; }
         public SelectList CategoryList { get; set; }
         public IEnumerable<int> CategoryIds { get; set; }
     }